+++ tag_database.hpp
@@ -7,7 +7,7 @@
 
 using TagSet = std::set<std::string>;
 using TagMap = std::unordered_map<std::string, TagSet>;
-using ChatMap = std::unordered_map<int32_t, TagMap>;
+using ChatMap = std::unordered_map<int64_t, TagMap>;
 
 class JsonTagDatabase{
 private:
@@ -16,23 +16,23 @@ private:
 public:
     JsonTagDatabase(const char *filename);
 
-    void CreateKeytag(int32_t chat_id, const std::string &keytag);
+    void CreateKeytag(int64_t chat_id, const std::string &keytag);
 
-    bool HasKeytag(int32_t chat_id, const std::string &keytag);
+    bool HasKeytag(int64_t chat_id, const std::string &keytag);
 
-    void DestroyKeytag(int32_t chat_id, const std::string &keytag);
+    void DestroyKeytag(int64_t chat_id, const std::string &keytag);
 
-    bool HasTag(int32_t chat_id, const std::string &keytag, const std::string &tag);
+    bool HasTag(int64_t chat_id, const std::string &keytag, const std::string &tag);
 
-    void AddTagsFor(int32_t chat_id, const std::string &keytag, const std::vector<std::string> &tags);
+    void AddTagsFor(int64_t chat_id, const std::string &keytag, const std::vector<std::string> &tags);
 
-    void RemoveTagsFor(int32_t chat_id, const std::string &keytag, const std::vector<std::string> &tags);
+    void RemoveTagsFor(int64_t chat_id, const std::string &keytag, const std::vector<std::string> &tags);
 
-    std::string GetTaggingMessage(int32_t chat_id, const std::string &keytag);
+    std::string GetTaggingMessage(int64_t chat_id, const std::string &keytag);
 
-    std::string GetKeytagsList(int32_t chat_id);
+    std::string GetKeytagsList(int64_t chat_id);
 
-    std::string GetTagsList(int32_t chat_id, const std::string &keytag);
+    std::string GetTagsList(int64_t chat_id, const std::string &keytag);
 private:
     void SaveToFile();
 };
